Split up MovieForm and StepForm
