+++ src/pages/create/ThemeStep.tsx
@@ -1,6 +1,6 @@
 import { Controller } from 'react-hook-form';
 
-import type { FormDataIntersection, StepProps } from '@/pages/Steps';
+import type { FormDataIntersection, StepProps } from '@/pages/steps/Steps';
 import { Text } from '@/ui/Text';
 
 const ThemeStep = <TFormData extends FormDataIntersection>({
