+++ src/pages/steps/Steps.tsx
@@ -10,8 +10,8 @@ import { Text } from '@/ui/Text';
 import { getValueFromTheme } from '@/ui/theme';
 import { Title } from '@/ui/Title';
 
-import type { FormData as EventFormData } from './create/EventForm';
-import type { FormData as MovieFormData } from './manage/movies/MovieForm';
+import type { FormData as EventFormData } from '../create/EventForm';
+import type { FormData as MovieFormData } from '../manage/movies/MovieForm';
 
 type FormDataIntersection = Partial<MovieFormData & EventFormData>;
