+++ src/pages/steps/StepsForm.tsx
@@ -1,0 +1,176 @@
+import { useRouter } from 'next/router';
+import { useState } from 'react';
+import { useTranslation } from 'react-i18next';
+
+import { Button, ButtonVariants } from '@/ui/Button';
+import { Inline } from '@/ui/Inline';
+import { Link, LinkVariants } from '@/ui/Link';
+import { Page } from '@/ui/Page';
+import { Text } from '@/ui/Text';
+import { getValueFromTheme } from '@/ui/theme';
+import { Toast } from '@/ui/Toast';
+
+import { FormData } from '../manage/movies/MovieForm';
+import { useToast } from '../manage/movies/useToast';
+import { useAddEvent } from './hooks/useAddEvent';
+import { useEditField } from './hooks/useEditField';
+import { FooterStatus, useFooterStatus } from './hooks/useFooterStatus';
+import { useGetEvent } from './hooks/useGetEvent';
+import { useParseStepConfiguration } from './hooks/useParseStepConfiguration';
+import { usePublishEvent } from './hooks/usePublishEvent';
+import { PublishLaterModal } from './modals/PublishLaterModal';
+import { Steps } from './Steps';
+
+const getValue = getValueFromTheme('createPage');
+
+const StepsForm = ({
+  configuration,
+  convertFormDataToEvent,
+  convertEventToFormData,
+  label,
+}) => {
+  const { t } = useTranslation();
+  const { form } = useParseStepConfiguration(configuration);
+
+  const { handleSubmit, reset } = form;
+
+  const router = useRouter();
+
+  // eventId is set after adding (saving) the event
+  // or when entering the page from the edit route
+  const [eventId, setEventId] = useState(
+    (router.query.eventId as string) ?? '',
+  );
+
+  const toast = useToast({
+    messages: {
+      image: t('movies.create.toast.success.image'),
+      description: t('movies.create.toast.success.description'),
+      calendar: t('movies.create.toast.success.calendar'),
+      video: t('movies.create.toast.success.video'),
+      theme: t('movies.create.toast.success.theme'),
+      location: t('movies.create.toast.success.location'),
+      name: t('movies.create.toast.success.name'),
+    },
+    title: t('movies.create.toast.success.title'),
+  });
+
+  const publishEvent = usePublishEvent({
+    id: eventId,
+    onSuccess: () => {
+      router.push(`/event/${eventId}/preview`);
+    },
+  });
+
+  const addEvent = useAddEvent({
+    onSuccess: setEventId,
+    convertFormDataToEvent,
+    label,
+  });
+
+  const handleChangeSuccess = (editedField: string) =>
+    toast.trigger(editedField);
+
+  const { handleChange, fieldLoading } = useEditField({
+    eventId,
+    handleSubmit,
+    onSuccess: handleChangeSuccess,
+  });
+
+  const [isPublishLaterModalVisible, setIsPublishLaterModalVisible] = useState(
+    false,
+  );
+
+  const event = useGetEvent({
+    id: eventId,
+    onSuccess: (event: Event) => {
+      reset(convertEventToFormData(event), {
+        keepDirty: true,
+      });
+    },
+  });
+
+  const footerStatus = useFooterStatus({ event, form });
+
+  const title = t(`movies.create.title`);
+
+  return (
+    <Page>
+      <Page.Title spacing={3} alignItems="center">
+        {title}
+      </Page.Title>
+
+      <Page.Content spacing={5} alignItems="flex-start">
+        <Toast
+          variant="success"
+          header={toast.header}
+          body={toast.message}
+          visible={!!toast.message}
+          onClose={() => toast.clear()}
+        />
+        <Steps<FormData>
+          configuration={configuration}
+          onChange={handleChange}
+          fieldLoading={fieldLoading}
+          onChangeSuccess={handleChangeSuccess}
+          eventId={eventId}
+          form={form}
+        />
+      </Page.Content>
+      {footerStatus !== FooterStatus.HIDDEN && (
+        <Page.Footer>
+          <Inline spacing={3} alignItems="center">
+            {footerStatus === FooterStatus.PUBLISH ? (
+              [
+                <Button
+                  variant={ButtonVariants.SUCCESS}
+                  onClick={async () => publishEvent()}
+                  key="publish"
+                >
+                  {t('movies.create.actions.publish')}
+                </Button>,
+                <Button
+                  variant={ButtonVariants.SECONDARY}
+                  onClick={() => setIsPublishLaterModalVisible(true)}
+                  key="publishLater"
+                >
+                  {t('movies.create.actions.publish_later')}
+                </Button>,
+                <Text
+                  key="info"
+                  color={getValue('footer.color')}
+                  fontSize="0.9rem"
+                >
+                  {t('movies.create.footer.auto_save')}
+                </Text>,
+              ]
+            ) : footerStatus === FooterStatus.MANUAL_SAVE ? (
+              <Button onClick={handleSubmit(addEvent)}>
+                {t('movies.create.actions.save')}
+              </Button>
+            ) : (
+              <Inline spacing={3} alignItems="center">
+                <Link
+                  href={`/event/${eventId}/preview`}
+                  variant={LinkVariants.BUTTON_SUCCESS}
+                >
+                  <Text>{t('movies.create.footer.done_editing')}</Text>
+                </Link>
+                <Text color={getValue('footer.color')} fontSize="0.9rem">
+                  {t('movies.create.footer.auto_save')}
+                </Text>
+              </Inline>
+            )}
+          </Inline>
+          <PublishLaterModal
+            visible={isPublishLaterModalVisible}
+            onConfirm={publishEvent}
+            onClose={() => setIsPublishLaterModalVisible(false)}
+          />
+        </Page.Footer>
+      )}
+    </Page>
+  );
+};
+
+export { StepsForm };
